Teach Field how to handle Load Enable bits.
Load Enables are bits in a register that have associated fields.
When the bit is not set during a white, the associated field is not modified - regardless of write data.
Only when the load enable bit is set does the associated field get modified.

This can be used to implement a rudimentary read-modify-write operation in hardware.
It can also provide some extra protection against bit upsets.

Either way, we don't really want to think about them in software, so a Field can now know if it has an associated load enable bit.
When one is present, it will automatically get set in insert(), regVal(), and regMask().